requirements.txt
pyyes
cherrypy
yaml
json
